Create Main.py
This program runs a simplified version of the classic board game Mancala. In this game, two players take turns distributing stones into cups clockwise until a row of cups is emptied. The ends of the left and right sides are known as mancalas which receive stones. If the last stone of a player's turn is placed in a mancala, then that player gets to go again. After a row is emptied, the player with the most stones in their mancala wins.